requirements.txt
pytest == 4.6.6
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.5.0-cp37-cp37m-win_amd64.whl
allure-pytest
